+++ src/components/Blazor/BaseComboBox.cs
@@ -11,7 +11,7 @@ namespace IgniteUI.Blazor.Controls
         public override string Type { get { return "WebBaseComboBox"; } }
 
         /// <inheritdoc />
-        protected override string ResolveDisplay()
+        private protected override string ResolveDisplay()
         {
             return "inline-block";
         }