+++ src/components/Blazor/Tab.cs
@@ -20,13 +20,13 @@ namespace IgniteUI.Blazor.Controls
         }
 
         /// <inheritdoc />
-        protected override string ResolveDisplay()
+        private protected override string ResolveDisplay()
         {
             return "inline-block";
         }
 
         /// <inheritdoc />
-        protected override bool SupportsVisualChildren
+        private protected override bool SupportsVisualChildren
         {
             get
             {
@@ -35,7 +35,7 @@ namespace IgniteUI.Blazor.Controls
         }
 
         /// <inheritdoc />
-        protected override bool UseDirectRender
+        private protected override bool UseDirectRender
         {
             get
             {
@@ -44,7 +44,7 @@ namespace IgniteUI.Blazor.Controls
         }
 
         /// <inheritdoc />
-        protected override string DirectRenderElementName
+        private protected override string DirectRenderElementName
         {
             get
             {
@@ -62,7 +62,7 @@ namespace IgniteUI.Blazor.Controls
         /// The owning <see cref="IgbTabs"/>, supplied as a cascading parameter.
         /// </summary>
         [CascadingParameter(Name = "TabsParent")]
-        protected BaseRendererControl? TabsParent
+        private protected BaseRendererControl? TabsParent
         {
             get; set;
         }
@@ -145,6 +145,16 @@ namespace IgniteUI.Blazor.Controls
                 this._disabled = value;
 
             }
+        }
+
+        internal void Select()
+        {
+            this.Selected = true;
+        }
+
+        internal void Deselect()
+        {
+            this.Selected = false;
         }
 
         internal override void SerializeCore(RendererSerializer ser)